requirements.txt
Django==5.2.4
Pillow==10.0.0
psycopg2-binary==2.9.7
